Get formatted text to display again after refactor
